+++ src/Route.php
@@ -5,6 +5,8 @@
  * @license   https://github.com/zendframework/zend-expressive-router/blob/master/LICENSE.md New BSD License
  */
 
+declare(strict_types=1);
+
 namespace Zend\Expressive\Router;
 
 use Fig\Http\Message\RequestMethodInterface as RequestMethod;
@@ -27,28 +29,28 @@ use Interop\Http\Server\MiddlewareInterface;
  */
 class Route
 {
-    const HTTP_METHOD_ANY = 0xff;
-    const HTTP_METHOD_SEPARATOR = ':';
+    public const HTTP_METHOD_ANY = null;
+    public const HTTP_METHOD_SEPARATOR = ':';
 
     /**
      * @var bool If HEAD was not provided to the Route instance, indicate
      *     support for the method is implicit.
      */
-    private $implicitHead = true;
+    private $implicitHead;
 
     /**
      * @var bool If OPTIONS was not provided to the Route instance, indicate
      *     support for the method is implicit.
      */
-    private $implicitOptions = true;
+    private $implicitOptions;
 
     /**
-     * @var int|string[] HTTP methods allowed with this route.
+     * @var null|string[] HTTP methods allowed with this route.
      */
-    private $methods = self::HTTP_METHOD_ANY;
+    private $methods;
 
     /**
-     * @var callable|string Middleware or service name of middleware associated with route.
+     * @var MiddlewareInterface Middleware associated with route.
      */
     private $middleware;
 
@@ -69,44 +71,22 @@ class Route
 
     /**
      * @param string $path Path to match.
-     * @param string|callable|MiddlewareInterface $middleware Middleware to use when this route is matched.
-     * @param int|array $methods Allowed HTTP methods; defaults to HTTP_METHOD_ANY.
-     * @param string|null $name the route name
-     * @throws Exception\InvalidArgumentException for invalid path type.
-     * @throws Exception\InvalidArgumentException for invalid middleware type.
-     * @throws Exception\InvalidArgumentException for any invalid HTTP method names.
+     * @param MiddlewareInterface $middleware Middleware to use when this route is matched.
+     * @param null|string[] $methods Allowed HTTP methods; defaults to HTTP_METHOD_ANY.
+     * @param null|string $name the route name
      */
-    public function __construct($path, $middleware, $methods = self::HTTP_METHOD_ANY, $name = null)
-    {
-        if (! is_string($path)) {
-            throw new Exception\InvalidArgumentException('Invalid path; must be a string');
-        }
-
-        if (! is_callable($middleware)
-            && ! $middleware instanceof MiddlewareInterface
-            && ! is_string($middleware)
-            && ! is_array($middleware)
-        ) {
-            throw new Exception\InvalidArgumentException(sprintf(
-                'Invalid middleware; must be a callable, a %s instance, a service name, '
-                . 'or an array of any of these types',
-                MiddlewareInterface::class
-            ));
-        }
-
-        if ($methods !== self::HTTP_METHOD_ANY && ! is_array($methods)) {
-            throw new Exception\InvalidArgumentException(sprintf(
-                'Invalid HTTP methods; must be an array or %s::HTTP_METHOD_ANY',
-                __CLASS__
-            ));
-        }
-
+    public function __construct(
+        string $path,
+        MiddlewareInterface $middleware,
+        array $methods = self::HTTP_METHOD_ANY,
+        string $name = null
+    ) {
         $this->path       = $path;
         $this->middleware = $middleware;
         $this->methods    = is_array($methods) ? $this->validateHttpMethods($methods) : $methods;
 
-        if (empty($name)) {
-            $name = ($this->methods === self::HTTP_METHOD_ANY)
+        if (! $name) {
+            $name = $this->methods === self::HTTP_METHOD_ANY
                 ? $path
                 : $path . '^' . implode(self::HTTP_METHOD_SEPARATOR, $this->methods);
         }
@@ -118,44 +98,33 @@ class Route
             && ! in_array(RequestMethod::METHOD_OPTIONS, $this->methods, true);
     }
 
-    /**
-     * @return string
-     */
-    public function getPath()
+    public function getPath() : string
     {
         return $this->path;
     }
 
     /**
      * Set the route name.
-     *
-     * @param string $name
      */
-    public function setName($name)
+    public function setName(string $name) : void
     {
-        $this->name = (string) $name;
+        $this->name = $name;
     }
 
-    /**
-     * @return string
-     */
-    public function getName()
+    public function getName() : string
     {
         return $this->name;
     }
 
-    /**
-     * @return string|callable|MiddlewareInterface
-     */
-    public function getMiddleware()
+    public function getMiddleware() : MiddlewareInterface
     {
         return $this->middleware;
     }
 
     /**
-     * @return int|string[] Returns HTTP_METHOD_ANY or array of allowed methods.
+     * @return null|string[] Returns HTTP_METHOD_ANY or array of allowed methods.
      */
-    public function getAllowedMethods()
+    public function getAllowedMethods() : ?array
     {
         return $this->methods;
     }
@@ -164,9 +133,8 @@ class Route
      * Indicate whether the specified method is allowed by the route.
      *
      * @param string $method HTTP method to test.
-     * @return bool
      */
-    public function allowsMethod($method)
+    public function allowsMethod(string $method) : bool
     {
         $method = strtoupper($method);
         if (RequestMethod::METHOD_HEAD === $method
@@ -180,38 +148,28 @@ class Route
         return false;
     }
 
-    /**
-     * @param array $options
-     */
-    public function setOptions(array $options)
+    public function setOptions(array $options) : void
     {
         $this->options = $options;
     }
 
-    /**
-     * @return array
-     */
-    public function getOptions()
+    public function getOptions() : array
     {
         return $this->options;
     }
 
     /**
      * Whether or not HEAD support is implicit (i.e., not explicitly specified)
-     *
-     * @return bool
      */
-    public function implicitHead()
+    public function implicitHead() : bool
     {
         return $this->implicitHead;
     }
 
     /**
      * Whether or not OPTIONS support is implicit (i.e., not explicitly specified)
-     *
-     * @return bool
      */
-    public function implicitOptions()
+    public function implicitOptions() : bool
     {
         return $this->implicitOptions;
     }
@@ -225,7 +183,7 @@ class Route
      * @return string[]
      * @throws Exception\InvalidArgumentException for any invalid method names.
      */
-    private function validateHttpMethods(array $methods)
+    private function validateHttpMethods(array $methods) : array
     {
         if (false === array_reduce($methods, function ($valid, $method) {
             if (false === $valid) {