+++ src/RouteResult.php
@@ -5,6 +5,8 @@
  * @license   https://github.com/zendframework/zend-expressive-router/blob/master/LICENSE.md New BSD License
  */
 
+declare(strict_types=1);
+
 namespace Zend\Expressive\Router;
 
 use Interop\Http\Server\MiddlewareInterface;
@@ -32,9 +34,9 @@ use Interop\Http\Server\MiddlewareInterface;
 class RouteResult
 {
     /**
-     * @var null|array
+     * @var array
      */
-    private $allowedMethods;
+    private $allowedMethods = [];
 
     /**
      * @var array
@@ -47,7 +49,7 @@ class RouteResult
     private $matchedRouteName;
 
     /**
-     * @var callable|string|MiddlewareInterface|array
+     * @var MiddlewareInterface
      */
     private $matchedMiddleware;
 
@@ -67,11 +69,9 @@ class RouteResult
     /**
      * Create an instance representing a route succes from the matching route.
      *
-     * @param Route $route
      * @param array $params Parameters associated with the matched route, if any.
-     * @return static
      */
-    public static function fromRoute(Route $route, array $params = [])
+    public static function fromRoute(Route $route, array $params = []) : self
     {
         $result                = new self();
         $result->success       = true;
@@ -83,10 +83,9 @@ class RouteResult
     /**
      * Create an instance representing a route failure.
      *
-     * @param null|int|array $methods HTTP methods allowed for the current URI, if any
-     * @return static
+     * @param null|array $methods HTTP methods allowed for the current URI, if any
      */
-    public static function fromRouteFailure($methods = null)
+    public static function fromRouteFailure(?array $methods = []) : self
     {
         $result = new self();
         $result->success = false;
@@ -104,10 +103,8 @@ class RouteResult
 
     /**
      * Does the result represent successful routing?
-     *
-     * @return bool
      */
-    public function isSuccess()
+    public function isSuccess() : bool
     {
         return $this->success;
     }
@@ -129,7 +126,7 @@ class RouteResult
      * If this result represents a failure, return false; otherwise, return the
      * matched route name.
      *
-     * @return string
+     * @return false|string
      */
     public function getMatchedRouteName()
     {
@@ -147,10 +144,8 @@ class RouteResult
     /**
      * Retrieve the matched middleware, if possible.
      *
-     * @return false|callable|string|MiddlewareInterface|array Returns false if
-     *     the result represents a failure; otherwise, a callable, a string
-     *     service name, a MiddlewareInterface instance, or array of any of
-     *     these types.
+     * @return false|MiddlewareInterface Returns false if the result
+     *     represents a failure; otherwise, a MiddlewareInterface instance.
      */
     public function getMatchedMiddleware()
     {
@@ -169,32 +164,26 @@ class RouteResult
      * Returns the matched params.
      *
      * Guaranted to return an array, even if it is simply empty.
-     *
-     * @return array
      */
-    public function getMatchedParams()
+    public function getMatchedParams() : array
     {
         return $this->matchedParams;
     }
 
     /**
      * Is this a routing failure result?
-     *
-     * @return bool
      */
-    public function isFailure()
+    public function isFailure() : bool
     {
         return (! $this->success);
     }
 
     /**
      * Does the result represent failure to route due to HTTP method?
-     *
-     * @return bool
      */
-    public function isMethodFailure()
+    public function isMethodFailure() : bool
     {
-        if ($this->isSuccess() || null === $this->allowedMethods) {
+        if ($this->isSuccess() || [] === $this->allowedMethods) {
             return false;
         }
 
@@ -206,16 +195,12 @@ class RouteResult
      *
      * @return string[] HTTP methods allowed
      */
-    public function getAllowedMethods()
+    public function getAllowedMethods() : array
     {
         if ($this->isSuccess()) {
             return $this->route
                 ? $this->route->getAllowedMethods()
                 : [];
-        }
-
-        if (null === $this->allowedMethods) {
-            return [];
         }
 
         return $this->allowedMethods;